Optimize React in production
This introduces two (key) production optimizations:
  - [inline elements
    transform](http://babeljs.io/docs/plugins/transform-react-inline-elements/):
Since React 0.14, `React.createElement` can be inlined by replacing
invocations with the equivalent ReactElements object. Hurray for less
function calls.
  - [constant elements
    transformer](http://babeljs.io/docs/plugins/transform-react-constant-elements/):

Also since React 0.14, ReactElements and their props are values (i.e.
elements with the same props are conceptually equivalent). As a result,
this optmization moves JSX with non-dynamic props out of the function
body so the exact same instance is returned for every render. Another hurray for
less allocations. As a bonus, this tells React not to bother diffing the
rest of the subtree, without having to set `shouldComponentUpdate`, for
an extra runtime performance boost.